feat: 블러 문구의 버튼에서도 팀 가입을 위해 팀 가입 모달 팀 대시보드로 위치 이동 및 TeamjoinModal로 정리
